Get fuel mix for one or more countries or regions

Get fuel mix for one or more countries or regions

get_fuel_mix(
  region_name,
  collapse_renewables = TRUE,
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name
A character vector with the names of one or more countries or regions to look up

collapse_renewables
Combine hydroelectricity and other renewables into a single category.

quiet
Suppress warnings if there is no data for that country or region.

region_code
Optional three-letter country or region codes to look up instead of the region_name

Value

A tibble of fuel mix for the countries or regions specified. That is, the number of quads of each fuel and the fraction of total primary energy coming from that fuel for each country or region:

region

The name of the country or region

year

The year reported

fuel

The name of the fuel

quads

The number of quads per year the country or region consumes

frac

The fraction of the country's energy that comes from that fuel

See also

Examples

get_fuel_mix("United States")

#> # A tibble: 5 x 5
#> # Groups:   region, year [1]
#>   region         year fuel        quads   frac
#>   <chr>         <int> <ord>       <dbl>  <dbl>
#> 1 United States  2019 Coal        10.7  0.120 
#> 2 United States  2019 Oil         35.1  0.391 
#> 3 United States  2019 Natural Gas 28.9  0.322 
#> 4 United States  2019 Nuclear      7.20 0.0802
#> 5 United States  2019 Renewables   7.82 0.0871

get_fuel_mix("World", collapse_renewables = FALSE)

#> # A tibble: 6 x 5
#>   region  year fuel        quads   frac
#>   <chr>  <int> <ord>       <dbl>  <dbl>
#> 1 World   2019 Coal        150.  0.270 
#> 2 World   2019 Oil         183.  0.331 
#> 3 World   2019 Natural Gas 134.  0.242 
#> 4 World   2019 Nuclear      23.6 0.0427
#> 5 World   2019 Hydro        35.7 0.0645
#> 6 World   2019 Renewables   27.5 0.0496

get_fuel_mix(region_code = "LCN")

#> # A tibble: 5 x 5
#> # Groups:   region, year [1]
#>   region                       year fuel         quads    frac
#>   <chr>                       <int> <ord>        <dbl>   <dbl>
#> 1 Latin America and Caribbean  2019 Coal         1.89  0.0548 
#> 2 Latin America and Caribbean  2019 Oil         14.4   0.417  
#> 3 Latin America and Caribbean  2019 Natural Gas  8.74  0.254  
#> 4 Latin America and Caribbean  2019 Nuclear      0.303 0.00880
#> 5 Latin America and Caribbean  2019 Renewables   9.15  0.266