Get top-down projections of Kaya variables for one or more countries or regions

Get top-down projections of Kaya variables for one or more countries or regions

get_top_down_values(region_name, quiet = FALSE, region_code = NULL)

Arguments

region_name

The name of a country or region to look up

quiet

Suppress warnings if there is no data for that country or region.

region_code

Optional three-letter country or region code to look up instead of the region_name

Value

a tibble of values for P, G, E, F, g, e, f, and ef

for each country or region:

region

The name of the country or region

P

Population, in billions

G

Gross domestic product, in trillions of constant 2015 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of constant 2015 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

See also

Examples

get_top_down_values("New Zealand")
#> # A tibble: 41 × 10
#>    region       year       P     G     g     E     F     e     f    ef
#>    <chr>       <int>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 New Zealand  2010 0.00432 0.158  36.6 0.746  35.5  4.71  47.7  224.
#>  2 New Zealand  2011 0.00438 0.163  37.1 0.786  34.6  4.83  44.0  213.
#>  3 New Zealand  2012 0.00445 0.169  37.9 0.797  35.0  4.73  44.0  208.
#>  4 New Zealand  2013 0.00452 0.172  38.1 0.797  34.4  4.63  43.2  200.
#>  5 New Zealand  2014 0.00458 0.177  38.6 0.804  34.6  4.55  43.0  195.
#>  6 New Zealand  2015 0.00465 0.182  39.0 0.810  34.9  4.46  43.2  192.
#>  7 New Zealand  2016 0.00473 0.187  39.5 0.836  35.9  4.48  42.9  192.
#>  8 New Zealand  2017 0.00481 0.191  39.8 0.885  37.6  4.63  42.4  196.
#>  9 New Zealand  2018 0.00489 0.197  40.3 0.912  38.0  4.63  41.7  193.
#> 10 New Zealand  2019 0.00496 0.202  40.7 0.933  38.1  4.61  40.9  188.
#> # … with 31 more rows
#> # ℹ Use `print(n = ...)` to see more rows
get_top_down_values("OECD")
#> # A tibble: 41 × 10
#>    region  year     P     G     g     E      F     e     f    ef
#>    <chr>  <int> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 OECD    2010  1.29  43.1  33.3  219. 12951.  5.09  59.0  301.
#>  2 OECD    2011  1.30  44.0  33.8  218. 12836.  4.95  59.0  292.
#>  3 OECD    2012  1.31  44.6  34.1  215. 12662.  4.83  58.8  284.
#>  4 OECD    2013  1.32  45.2  34.4  218. 12774.  4.81  58.7  282.
#>  5 OECD    2014  1.33  46.2  34.9  217. 12641.  4.68  58.4  273.
#>  6 OECD    2015  1.33  47.4  35.5  217. 12541.  4.57  57.8  265.
#>  7 OECD    2016  1.34  48.3  35.9  218. 12485.  4.52  57.3  259.
#>  8 OECD    2017  1.35  49.5  36.7  224. 12347.  4.52  55.2  249.
#>  9 OECD    2018  1.36  50.6  37.3  226. 12293.  4.47  54.4  243.
#> 10 OECD    2019  1.37  51.6  37.8  226. 12107.  4.39  53.5  235.
#> # … with 31 more rows
#> # ℹ Use `print(n = ...)` to see more rows
get_top_down_values(region_code = "PAK")
#> # A tibble: 41 × 10
#>    region    year     P     G     g     E     F     e     f    ef
#>    <chr>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Pakistan  2010 0.189 0.216  1.14  2.58  151.  11.9  58.4  697.
#>  2 Pakistan  2011 0.192 0.226  1.18  2.61  153.  11.5  58.6  677.
#>  3 Pakistan  2012 0.195 0.238  1.22  2.69  156.  11.3  58.1  657.
#>  4 Pakistan  2013 0.197 0.249  1.26  2.75  166.  11.1  60.4  668.
#>  5 Pakistan  2014 0.200 0.260  1.30  2.84  172.  10.9  60.6  663.
#>  6 Pakistan  2015 0.203 0.271  1.34  2.92  178.  10.7  61.1  656.
#>  7 Pakistan  2016 0.205 0.284  1.38  3.06  185.  10.8  60.5  653.
#>  8 Pakistan  2017 0.208 0.298  1.43  3.21  189.  10.8  58.7  633.
#>  9 Pakistan  2018 0.210 0.313  1.49  3.31  195.  10.6  58.8  621.
#> 10 Pakistan  2019 0.213 0.328  1.54  3.37  199.  10.3  58.9  606.
#> # … with 31 more rows
#> # ℹ Use `print(n = ...)` to see more rows