Kaya identity data for many countries

A dataset containing Kaya identity parameters P, G, E, F, g, e, f, and ef for many countries

kaya_data

Format

A tibble containing 5133 rows and 14 variables:

region

Country or region name

region_code

Three-letter country or region code

geography

Geographic category: "nation", "region", or "world"

year

The year

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of tons

g

Per-capita GDP, in thousands of constant 2010 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

G_ppp

Gross domestic product adjusted for purchasing power parity, in trillions of constant 2011 international dollars

G_mer

Gross domestic product at market-exchange-rate, in trillions of constant 2010 U.S. dollars

Source

https://data.worldbank.org/indicator/SP.POP.TOTL, https://data.worldbank.org/indicator/NY.GDP.MKTP.KD, and https://www.bp.com/en/global/corporate/energy-economics/statistical-review-of-world-energy/downloads.html