Get power output from generation sources

Nameplate capacity and capacity factors for different electrical generation technologies. The average power supplied over a year is the nameplate capacity times the capacity factor.

generation_capacity()

Value

a tibble of values for generation sources

fuel

Energy source: Coal, Nuclear, Gas, Solar Thermal, Solar Photovoltaic, Onshore Wind, or Offshore Wind

description

Text description of the power source

nameplate_capacity

Maximum sustained power output, in megawatts

capacity_factor

Capacity factor: the fraction of the nameplate capacity that the plant can provide, averaged over a typical year

Details

Data for fossil fuels comes from EIA

References

Environmental Protection Agency (2018) "Electric Power Monthly," (October, 2018) https://www.eia.gov/electricity/monthly/archive/october2018.pdf, Table 6.7.A.

Pielke, Jr., Roger A., The Climate Fix (Basic Books, 2010).

Examples

gc <- generation_capacity()
gc
#> # A tibble: 7 × 4
#>   fuel               description                            nameplate_…¹ capac…²
#>   <chr>              <chr>                                         <dbl>   <dbl>
#> 1 Coal               Large coal-fired power plant                   1000    0.53
#> 2 Nuclear            Large nuclear power plant                      1000    0.75
#> 3 Natural Gas        Gas-fired power plant                           500    0.56
#> 4 Photovoltaic Solar Concentrated solar-thermal power plant          100    0.25
#> 5 Solar Thermal      Photovoltaic solar farm                         100    0.25
#> 6 Onshore Wind       Onshore Wind turbine                              6    0.42
#> 7 Offshore Wind      Offshore Wind turbine                            13    0.5 
#> # … with abbreviated variable names ¹​nameplate_capacity, ²​capacity_factor