Get fuel mix for one or more countries or regions

Get fuel mix for one or more countries or regions

get_fuel_mix(
  region_name,
  collapse_renewables = TRUE,
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name

A character vector with the names of one or more countries or regions to look up

collapse_renewables

Combine hydroelectricity and other renewables into a single category.

quiet

Suppress warnings if there is no data for that country or region.

region_code

Optional three-letter country or region codes to look up instead of the region_name

Value

A tibble of fuel mix for the countries or regions specified. That is, the number of quads of each fuel and the fraction of total primary energy coming from that fuel for each country or region:

region

The name of the country or region

year

The year reported

fuel

The name of the fuel

quads

The number of quads per year the country or region consumes

frac

The fraction of the country's energy that comes from that fuel

See also

Examples

get_fuel_mix("United States")
#> # A tibble: 5 × 5
#> # Groups:   region, year [1]
#>   region         year fuel        quads   frac
#>   <chr>         <int> <ord>       <dbl>  <dbl>
#> 1 United States  2021 Coal        10.0  0.114 
#> 2 United States  2021 Oil         33.5  0.380 
#> 3 United States  2021 Natural Gas 28.2  0.320 
#> 4 United States  2021 Nuclear      7.02 0.0796
#> 5 United States  2021 Renewables   9.39 0.107 
get_fuel_mix("World", collapse_renewables = FALSE)
#> # A tibble: 6 × 5
#>   region  year fuel        quads   frac
#>   <chr>  <int> <ord>       <dbl>  <dbl>
#> 1 World   2021 Coal        152.  0.269 
#> 2 World   2021 Oil         175.  0.310 
#> 3 World   2021 Natural Gas 138.  0.244 
#> 4 World   2021 Nuclear      24.0 0.0425
#> 5 World   2021 Hydro        38.2 0.0676
#> 6 World   2021 Renewables   37.8 0.0671
get_fuel_mix(region_code = "LCN")
#> # A tibble: 5 × 5
#> # Groups:   region, year [1]
#>   region                       year fuel         quads    frac
#>   <chr>                       <int> <ord>        <dbl>   <dbl>
#> 1 Latin America and Caribbean  2021 Coal         1.60  0.0480 
#> 2 Latin America and Caribbean  2021 Oil         13.1   0.394  
#> 3 Latin America and Caribbean  2021 Natural Gas  8.58  0.257  
#> 4 Latin America and Caribbean  2021 Nuclear      0.321 0.00961
#> 5 Latin America and Caribbean  2021 Renewables   9.75  0.292