Get fuel mix for a country or region

Get fuel mix for a country or region

get_fuel_mix(
  region_name,
  collapse_renewables = TRUE,
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name
The name of a country or region to look up

collapse_renewables
Combine hydroelectricity and other renewables into a single category.

quiet
Suppress warnings if there is no data for that country or region.

region_code
Optional three-letter country or region code to look up instead of the `region_name`

Value

a tibble of fuel mix for the country or region. That is, the number of quads of each fuel and the fraction of total primary energy coming from that fuel.

Examples

get_fuel_mix("United States")

#> # A tibble: 5 x 5
#>   region         year fuel        quads   frac
#>   <chr>         <int> <ord>       <dbl>  <dbl>
#> 1 United States  2018 Coal        12.6  0.138 
#> 2 United States  2018 Oil         36.5  0.400 
#> 3 United States  2018 Natural Gas 27.9  0.305 
#> 4 United States  2018 Nuclear      7.63 0.0836
#> 5 United States  2018 Renewables   6.71 0.0735

get_fuel_mix("World", collapse_renewables = FALSE)

#> # A tibble: 6 x 5
#>   region  year fuel        quads   frac
#>   <chr>  <int> <ord>       <dbl>  <dbl>
#> 1 World   2018 Coal        150.  0.272 
#> 2 World   2018 Oil         185.  0.336 
#> 3 World   2018 Natural Gas 131.  0.239 
#> 4 World   2018 Nuclear      24.3 0.0441
#> 5 World   2018 Hydro        37.7 0.0684
#> 6 World   2018 Renewables   22.3 0.0405

get_fuel_mix(region_code = "LCN")

#> # A tibble: 5 x 5
#>   region                       year fuel         quads    frac
#>   <chr>                       <int> <ord>        <dbl>   <dbl>
#> 1 Latin America and Caribbean  2018 Coal         1.90  0.0539 
#> 2 Latin America and Caribbean  2018 Oil         15.8   0.448  
#> 3 Latin America and Caribbean  2018 Natural Gas  8.80  0.249  
#> 4 Latin America and Caribbean  2018 Nuclear      0.324 0.00919
#> 5 Latin America and Caribbean  2018 Renewables   8.45  0.240