Get fuel mix for one or more countries or regions

Get fuel mix for one or more countries or regions

get_fuel_mix(
  region_name,
  collapse_renewables = TRUE,
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name

A character vector with the names of one or more countries or regions to look up

collapse_renewables

Combine hydroelectricity and other renewables into a single category.

quiet

Suppress warnings if there is no data for that country or region.

region_code

Optional three-letter country or region codes to look up instead of the region_name

Value

A tibble of fuel mix for the countries or regions specified. That is, the number of quads of each fuel and the fraction of total primary energy coming from that fuel for each country or region:

region

The name of the country or region

year

The year reported

fuel

The name of the fuel

quads

The number of quads per year the country or region consumes

frac

The fraction of the country's energy that comes from that fuel

Note

In the latest data from the Energy Institute, there are small discrepancies between the sums of energy for each fuel and the totals, in both quads and frac, for Hong Kong and Sri Lanka, as described in the documentation for fuel_mix.

See also

Examples

get_fuel_mix("United States")
#> # A tibble: 5 × 5
#> # Groups:   region, year [1]
#>   region         year fuel        quads   frac
#>   <chr>         <int> <ord>       <dbl>  <dbl>
#> 1 United States  2022 Coal         9.35 0.103 
#> 2 United States  2022 Oil         34.3  0.377 
#> 3 United States  2022 Natural Gas 30.1  0.331 
#> 4 United States  2022 Nuclear      6.93 0.0762
#> 5 United States  2022 Renewables  10.3  0.113 
get_fuel_mix("World", collapse_renewables = FALSE)
#> # A tibble: 6 × 5
#>   region  year fuel        quads   frac
#>   <chr>  <int> <ord>       <dbl>  <dbl>
#> 1 World   2022 Coal        153.  0.267 
#> 2 World   2022 Oil         181.  0.316 
#> 3 World   2022 Natural Gas 134.  0.235 
#> 4 World   2022 Nuclear      22.9 0.0399
#> 5 World   2022 Hydro        38.6 0.0673
#> 6 World   2022 Renewables   42.8 0.0748
get_fuel_mix(region_code = "LCN")
#> # A tibble: 5 × 5
#> # Groups:   region, year [1]
#>   region                       year fuel         quads    frac
#>   <chr>                       <int> <ord>        <dbl>   <dbl>
#> 1 Latin America and Caribbean  2022 Coal         1.36  0.0371 
#> 2 Latin America and Caribbean  2022 Oil         15.6   0.425  
#> 3 Latin America and Caribbean  2022 Natural Gas  8.81  0.239  
#> 4 Latin America and Caribbean  2022 Nuclear      0.284 0.00772
#> 5 Latin America and Caribbean  2022 Renewables  10.7   0.291