Get Kaya data for one or more countries or regions

Get Kaya data for one or more countries or regions

get_kaya_data(
  region_name,
  gdp = c("MER", "PPP"),
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name

The name of one or more countries or regions to look up

gdp

Use market exchange rates (MER) or purchasing power parity (PPP). Default is MER.

quiet

Suppress warnings if there is no such country or region.

region_code

Optional three-letter country or region codes to look up instead of the region_name

Value

a tibble of Kaya identity data for the countries or regions specified:

region

The name of the country or region

year

The year

P

Population, in billions

G

Gross domestic product, in trillions of constant 2015 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

Details

Units for G, g, e, and ef depend on whether the data is requested in MER or PPP dollars: For MER, dollars are constant 2015 U.S. dollars. For PPP, dollars are constant 2017 international dollars.

     _P_ and MER values for GDP and related quantities are available
     from 1960 onward.

     PPP values for GDP and related quantities are only available from
     1990 onward.

     Energy-related values (_E_, _F_, and derived quantities) are
     available from 1965 onward.

     Note that emissions (_F_, _f_, and _ef_) are reported as millions
     of metric tons of carbon dioxide, not carbon.

See also

Examples

get_kaya_data("Brazil")
#> # A tibble: 63 × 10
#>    region  year     P     G     E     F     g     e     f    ef
#>    <ord>  <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Brazil  2022 0.215  1.90  12.7  443.  8.83  6.68  34.9  233.
#>  2 Brazil  2021 0.214  1.85  12.2  462.  8.62  6.59  37.9  250.
#>  3 Brazil  2020 0.213  1.76  11.6  408.  8.26  6.58  35.3  232.
#>  4 Brazil  2019 0.212  1.82  12.1  436.  8.59  6.63  36.1  239.
#>  5 Brazil  2018 0.210  1.80  11.9  442.  8.55  6.60  37.3  246.
#>  6 Brazil  2017 0.209  1.77  11.8  462.  8.47  6.69  39.1  261.
#>  7 Brazil  2016 0.207  1.74  11.7  456.  8.43  6.72  38.9  262.
#>  8 Brazil  2015 0.205  1.80  12.0  495.  8.78  6.66  41.2  275.
#>  9 Brazil  2014 0.203  1.87  12.2  512.  9.18  6.51  42.1  274.
#> 10 Brazil  2013 0.202  1.86  11.9  487   9.22  6.38  41.1  262.
#> # ℹ 53 more rows
get_kaya_data("United Kingdom", "PPP")
#> # A tibble: 63 × 10
#>    region          year      P     G     E     F     g     e     f    ef
#>    <ord>          <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United Kingdom  2022 0.0670  3.14  6.93  345.  46.8  2.21  49.7  110.
#>  2 United Kingdom  2021 0.0670  3.01  6.82  341.  44.9  2.27  50.0  113.
#>  3 United Kingdom  2020 0.0671  2.80  6.73  321.  41.7  2.40  47.7  115.
#>  4 United Kingdom  2019 0.0668  3.15  7.50  381.  47.1  2.38  50.9  121.
#>  5 United Kingdom  2018 0.0665  3.10  7.66  396.  46.6  2.47  51.8  128.
#>  6 United Kingdom  2017 0.0661  3.05  7.69  404.  46.1  2.52  52.6  133.
#>  7 United Kingdom  2016 0.0656  2.97  7.69  415.  45.3  2.59  54.0  140.
#>  8 United Kingdom  2015 0.0651  2.91  7.76  439.  44.7  2.67  56.6  151.
#>  9 United Kingdom  2014 0.0646  2.84  7.64  456.  44.0  2.69  59.7  161.
#> 10 United Kingdom  2013 0.0641  2.75  8.09  498.  42.9  2.94  61.5  181.
#> # ℹ 53 more rows
get_kaya_data(region_name = "United States")
#> # A tibble: 63 × 10
#>    region         year     P     G     E     F     g     e     f    ef
#>    <ord>         <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United States  2022 0.333  21.0  90.9 4826.  62.9  4.34  53.1  230.
#>  2 United States  2021 0.332  20.5  88.5 4768.  61.8  4.31  53.9  232.
#>  3 United States  2020 0.332  19.4  83.9 4463.  58.5  4.33  53.2  230.
#>  4 United States  2019 0.328  19.9  90.7 4982.  60.7  4.55  54.9  250.
#>  5 United States  2018 0.327  19.5  91.3 5133.  59.6  4.69  56.2  263.
#>  6 United States  2017 0.325  18.9  88.1 4979.  58.2  4.66  56.5  263.
#>  7 United States  2016 0.323  18.5  87.8 5038   57.3  4.74  57.4  272.
#>  8 United States  2015 0.321  18.2  87.9 5138.  56.8  4.83  58.5  282.
#>  9 United States  2014 0.318  17.7  88.7 5252.  55.7  5.00  59.2  296.
#> 10 United States  2013 0.316  17.3  87.8 5247.  54.8  5.07  59.8  303.
#> # ℹ 53 more rows
get_kaya_data(region_code = "MYS")
#> # A tibble: 63 × 10
#>    region    year      P     G     E     F     g     e     f    ef
#>    <ord>    <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Malaysia  2022 0.0339 0.386  4.59  273. 11.4   11.9  59.5  707.
#>  2 Malaysia  2021 0.0336 0.355  4.34  256. 10.6   12.2  58.9  720.
#>  3 Malaysia  2020 0.0332 0.344  4.08  247  10.4   11.8  60.6  717.
#>  4 Malaysia  2019 0.0328 0.365  4.24  256. 11.1   11.6  60.4  702.
#>  5 Malaysia  2018 0.0324 0.349  4.12  251. 10.8   11.8  60.9  719.
#>  6 Malaysia  2017 0.0320 0.333  4.06  241. 10.4   12.2  59.5  724.
#>  7 Malaysia  2016 0.0315 0.315  4.00  253.  9.98  12.7  63.2  803.
#>  8 Malaysia  2015 0.0311 0.301  3.80  247.  9.70  12.6  65.0  820.
#>  9 Malaysia  2014 0.0306 0.287  3.74  243.  9.37  13.1  65.0  848.
#> 10 Malaysia  2013 0.0301 0.271  3.71  234.  8.98  13.7  63.2  865.
#> # ℹ 53 more rows