Get Kaya data for one or more countries or regions

Get Kaya data for one or more countries or regions

get_kaya_data(
  region_name,
  gdp = c("MER", "PPP"),
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name
The name of one or more countries or regions to look up

gdp
Use market exchange rates (MER) or purchasing power parity (PPP). Default is MER.

quiet
Suppress warnings if there is no such country or region.

region_code
Optional three-letter country or region codes to look up instead of the region_name

Value

a tibble of Kaya identity data for the countries or regions specified:

region

The name of the country or region

year

The year

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

Details

Units for G, g, e, and ef depend on whether the data is requested in MER or PPP dollars: For MER, dollars are constant 2010 U.S. dollars. For PPP, dollars are constant 2011 international dollars.

     _P_ and MER values for GDP and related quantities are available
     from 1960 onward.

     PPP values for GDP and related quantities are only available from
     1990 onward.

     Energy-related values (_E_, _F_, and derived quantities) are
     available from 1965 onward.

     Note that emissions (_F_, _f_, and _ef_) are reported as millions
     of metric tons of carbon dioxide, not carbon.

See also

Examples

get_kaya_data("Brazil")

#> # A tibble: 61 × 10
#>    region  year     P     G     E     F     g     e     f    ef
#>    <ord>  <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Brazil  2020 0.213  2.27  11.4  418.  10.7  5.02  36.7  184.
#>  2 Brazil  2019 0.211  2.36  11.8  445.  11.2  4.98  37.8  188.
#>  3 Brazil  2018 0.209  2.33  11.5  446.  11.1  4.93  38.8  191.
#>  4 Brazil  2017 0.208  2.29  11.4  461.  11.0  4.98  40.4  201.
#>  5 Brazil  2016 0.206  2.26  11.3  454.  11.0  4.98  40.3  201.
#>  6 Brazil  2015 0.204  2.34  11.6  491.  11.4  4.95  42.4  210.
#>  7 Brazil  2014 0.203  2.42  11.7  508.  12.0  4.84  43.3  210.
#>  8 Brazil  2013 0.201  2.41  11.5  486.  12.0  4.76  42.4  201.
#>  9 Brazil  2012 0.199  2.34  11.1  446.  11.7  4.72  40.3  190.
#> 10 Brazil  2011 0.198  2.30  10.9  427.  11.6  4.73  39.3  186.
#> # … with 51 more rows

get_kaya_data("United Kingdom", "PPP")

#> # A tibble: 61 × 10
#>    region          year      P     G     E     F     g     e     f    ef
#>    <ord>          <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United Kingdom  2020 0.0672  2.80  6.53  319.  41.6  2.34  48.9  114.
#>  2 United Kingdom  2019 0.0668  3.10  7.32  380.  46.4  2.36  51.9  123.
#>  3 United Kingdom  2018 0.0665  3.06  7.53  398.  46.0  2.46  52.8  130.
#>  4 United Kingdom  2017 0.0661  3.02  7.54  404.  45.7  2.50  53.6  134.
#>  5 United Kingdom  2016 0.0656  2.97  7.57  416.  45.3  2.55  55.0  140.
#>  6 United Kingdom  2015 0.0651  2.92  7.66  440.  44.8  2.62  57.5  151.
#>  7 United Kingdom  2014 0.0646  2.85  7.57  459.  44.2  2.66  60.6  161.
#>  8 United Kingdom  2013 0.0641  2.77  8.04  501.  43.2  2.90  62.3  181.
#>  9 United Kingdom  2012 0.0637  2.71  8.08  513.  42.6  2.98  63.4  189.
#> 10 United Kingdom  2011 0.0633  2.68  7.99  495.  42.3  2.99  62.0  185.
#> # … with 51 more rows

get_kaya_data(region_name = "United States")

#> # A tibble: 61 × 10
#>    region         year     P     G     E     F     g     e     f    ef
#>    <ord>         <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United States  2020 0.329  17.7  83.2 4457.  53.7  4.70  53.6  252.
#>  2 United States  2019 0.328  18.3  89.9 5029.  55.9  4.90  55.9  274.
#>  3 United States  2018 0.327  18.0  90.6 5166.  55.0  5.05  57.0  288.
#>  4 United States  2017 0.325  17.4  87.4 5003.  53.6  5.01  57.2  287.
#>  5 United States  2016 0.323  17.0  87.2 5061.  52.7  5.11  58.1  297.
#>  6 United States  2015 0.321  16.8  87.3 5166.  52.2  5.21  59.2  308.
#>  7 United States  2014 0.318  16.3  88.1 5278.  51.1  5.42  59.9  325.
#>  8 United States  2013 0.316  15.9  87.2 5268.  50.2  5.50  60.4  332.
#>  9 United States  2012 0.314  15.6  84.9 5102.  49.6  5.46  60.1  328.
#> 10 United States  2011 0.312  15.2  87.2 5348.  48.9  5.73  61.3  351.
#> # … with 51 more rows

get_kaya_data(region_code = "MYS")

#> # A tibble: 61 × 10
#>    region    year      P     G     E     F     g     e     f    ef
#>    <ord>    <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Malaysia  2020 0.0324 0.377  3.90  256. 11.6   10.3  65.7  680.
#>  2 Malaysia  2019 0.0319 0.399  4.20  263. 12.5   10.5  62.6  659.
#>  3 Malaysia  2018 0.0315 0.382  4.10  255. 12.1   10.7  62.2  667.
#>  4 Malaysia  2017 0.0311 0.365  4.05  247. 11.7   11.1  61.1  678.
#>  5 Malaysia  2016 0.0307 0.345  3.99  258. 11.2   11.6  64.8  749.
#>  6 Malaysia  2015 0.0303 0.330  3.79  254. 10.9   11.5  67.0  768.
#>  7 Malaysia  2014 0.0299 0.314  3.73  249. 10.5   11.9  66.8  793.
#>  8 Malaysia  2013 0.0295 0.297  3.70  240. 10.1   12.5  64.7  808.
#>  9 Malaysia  2012 0.0291 0.283  3.53  229.  9.74  12.5  64.9  810.
#> 10 Malaysia  2011 0.0287 0.269  3.29  216.  9.37  12.3  65.7  805.
#> # … with 51 more rows