Get Kaya data for a country or region

Get Kaya data for a country or region

get_kaya_data(
  region_name,
  gdp = c("MER", "PPP"),
  quiet = FALSE,
  region_code = NULL
)

Arguments

region_name
The name of a country or region to look up

gdp
Use market exchange rates (MER) or purchasing power parity (PPP). Default is MER.

quiet
Suppress warnings if there is no such country or region.

region_code
Optional three-letter country or region code to look up instead of the region_name

Value

a tibble of Kaya identity data for the country or region:

region

The name of the country or region

year

The year

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

Details

Units for G, g, e, and ef depend on whether the data is requested in MER or PPP dollars: For MER, dollars are constant 2010 U.S. dollars. For PPP, dollars are constant 2011 international dollars.

     P and MER values for GDP and related quantities are available from
     1960 onward.

     PPP values for GDP and related quantities are only available from
     1990 onward.

     Energy-related values (E, F, and derived quantities) are
     available from 1965 onward.

     Note that emissions (F, f, and ef) use millions of metric tons of
     carbon dioxide, not carbon.

Examples

get_kaya_data("Brazil")

#> # A tibble: 59 x 10
#>    region  year      P     G      E     F     g     e     f    ef
#>    <ord>  <int>  <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Brazil  1960 0.0722 0.247 NA      NA    3.42 NA     NA     NA 
#>  2 Brazil  1961 0.0743 0.272 NA      NA    3.66 NA     NA     NA 
#>  3 Brazil  1962 0.0765 0.286 NA      NA    3.74 NA     NA     NA 
#>  4 Brazil  1963 0.0788 0.289 NA      NA    3.66 NA     NA     NA 
#>  5 Brazil  1964 0.0811 0.299 NA      NA    3.69 NA     NA     NA 
#>  6 Brazil  1965 0.0834 0.308  0.892  51.5  3.69  2.90  57.7  167.
#>  7 Brazil  1966 0.0857 0.321  0.973  56.1  3.74  3.04  57.6  175.
#>  8 Brazil  1967 0.0880 0.336  1.01   58.0  3.82  3.00  57.5  172.
#>  9 Brazil  1968 0.0904 0.375  1.16   68.1  4.15  3.08  58.9  182.
#> 10 Brazil  1969 0.0927 0.411  1.26   74.3  4.44  3.07  58.8  181.
#> # … with 49 more rows

get_kaya_data("United Kingdom", "PPP")

#> # A tibble: 59 x 10
#>    region          year      P     G     E     F     g     e     f    ef
#>    <ord>          <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United Kingdom  1960 0.0524    NA NA      NA     NA    NA  NA      NA
#>  2 United Kingdom  1961 0.0528    NA NA      NA     NA    NA  NA      NA
#>  3 United Kingdom  1962 0.0532    NA NA      NA     NA    NA  NA      NA
#>  4 United Kingdom  1963 0.0536    NA NA      NA     NA    NA  NA      NA
#>  5 United Kingdom  1964 0.054     NA NA      NA     NA    NA  NA      NA
#>  6 United Kingdom  1965 0.0543    NA  7.89  688.    NA    NA  87.2    NA
#>  7 United Kingdom  1966 0.0546    NA  7.92  681.    NA    NA  85.9    NA
#>  8 United Kingdom  1967 0.0549    NA  7.92  670.    NA    NA  84.5    NA
#>  9 United Kingdom  1968 0.0552    NA  8.21  690.    NA    NA  84.0    NA
#> 10 United Kingdom  1969 0.0554    NA  8.50  707.    NA    NA  83.2    NA
#> # … with 49 more rows

get_kaya_data(region_name = "United States")

#> # A tibble: 59 x 10
#>    region         year     P     G     E     F     g     e     f    ef
#>    <ord>         <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 United States  1960 0.181  3.17  NA     NA   17.6  NA    NA     NA 
#>  2 United States  1961 0.184  3.24  NA     NA   17.7  NA    NA     NA 
#>  3 United States  1962 0.187  3.44  NA     NA   18.5  NA    NA     NA 
#>  4 United States  1963 0.189  3.59  NA     NA   19.0  NA    NA     NA 
#>  5 United States  1964 0.192  3.80  NA     NA   19.8  NA    NA     NA 
#>  6 United States  1965 0.194  4.04  49.6 3480.  20.8  12.3  70.2  860.
#>  7 United States  1966 0.197  4.31  52.4 3675.  21.9  12.2  70.2  853.
#>  8 United States  1967 0.199  4.42  54.2 3773.  22.2  12.3  69.6  854.
#>  9 United States  1968 0.201  4.63  57.5 3994.  23.1  12.4  69.5  863.
#> 10 United States  1969 0.203  4.77  60.6 4170.  23.5  12.7  68.9  874.
#> # … with 49 more rows

get_kaya_data(region_code = "MYS")

#> # A tibble: 59 x 10
#>    region    year       P      G       E     F     g     e     f    ef
#>    <ord>    <int>   <dbl>  <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Malaysia  1960 0.00816 0.0110 NA      NA     1.35 NA     NA     NA 
#>  2 Malaysia  1961 0.00842 0.0119 NA      NA     1.41 NA     NA     NA 
#>  3 Malaysia  1962 0.00869 0.0126 NA      NA     1.45 NA     NA     NA 
#>  4 Malaysia  1963 0.00897 0.0136 NA      NA     1.51 NA     NA     NA 
#>  5 Malaysia  1964 0.00925 0.0143 NA      NA     1.55 NA     NA     NA 
#>  6 Malaysia  1965 0.00953 0.0154  0.0985  7.02  1.62  6.40  71.2  456.
#>  7 Malaysia  1966 0.00979 0.0166  0.115   8.18  1.70  6.92  71.2  493.
#>  8 Malaysia  1967 0.0100  0.0172  0.116   8.29  1.72  6.74  71.4  481.
#>  9 Malaysia  1968 0.0103  0.0186  0.118   8.39  1.81  6.36  70.9  451.
#> 10 Malaysia  1969 0.0105  0.0195  0.124   8.67  1.85  6.35  69.9  444.
#> # … with 49 more rows