Get top-down projections of Kaya variables for one or more countries or regions

Get top-down projections of Kaya variables for one or more countries or regions

get_top_down_values(region_name, quiet = FALSE, region_code = NULL)

Arguments

region_name

The name of a country or region to look up

quiet

Suppress warnings if there is no data for that country or region.

region_code

Optional three-letter country or region code to look up instead of the region_name

Value

a tibble of values for P, G, E, F, g, e, f, and ef for each country or region:

region

The name of the country or region

P

Population, in billions

G

Gross domestic product, in trillions of constant 2015 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of constant 2015 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

See also

Examples

get_top_down_values("New Zealand")
#> # A tibble: 41 × 10
#>    region       year       P     G     g     E     F     e     f    ef
#>    <chr>       <int>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 New Zealand  2010 0.00432 0.158  36.6 0.747  34.7  4.73  46.4  219.
#>  2 New Zealand  2011 0.00438 0.162  37.0 0.787  33.8  4.85  42.9  208.
#>  3 New Zealand  2012 0.00445 0.168  37.8 0.798  34.2  4.74  42.9  203.
#>  4 New Zealand  2013 0.00452 0.172  38.0 0.799  33.6  4.65  42.0  195.
#>  5 New Zealand  2014 0.00458 0.177  38.5 0.805  33.7  4.56  41.9  191.
#>  6 New Zealand  2015 0.00465 0.181  38.9 0.811  34.1  4.48  42.0  188.
#>  7 New Zealand  2016 0.00473 0.187  39.4 0.838  35.0  4.49  41.8  188.
#>  8 New Zealand  2017 0.00481 0.191  39.7 0.887  36.7  4.64  41.4  192.
#>  9 New Zealand  2018 0.00489 0.197  40.2 0.913  37.1  4.64  40.6  188.
#> 10 New Zealand  2019 0.00496 0.202  40.7 0.935  37.2  4.63  39.8  184.
#> # ℹ 31 more rows
get_top_down_values("OECD")
#> # A tibble: 41 × 10
#>    region  year     P     G     g     E      F     e     f    ef
#>    <chr>  <int> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 OECD    2010  1.29  43.2  33.5  221. 13047.  5.10  59.2  302.
#>  2 OECD    2011  1.30  44.1  34.0  219. 12932.  4.96  59.1  293.
#>  3 OECD    2012  1.30  44.7  34.3  216. 12756.  4.84  58.9  285.
#>  4 OECD    2013  1.31  45.4  34.6  219. 12869.  4.82  58.8  284.
#>  5 OECD    2014  1.32  46.4  35.1  218. 12735.  4.69  58.5  274.
#>  6 OECD    2015  1.33  47.6  35.7  218. 12635.  4.58  58.0  266.
#>  7 OECD    2016  1.34  48.4  36.2  219. 12578.  4.53  57.4  260.
#>  8 OECD    2017  1.35  49.7  36.9  225. 12438.  4.53  55.3  250.
#>  9 OECD    2018  1.35  50.8  37.5  227. 12384.  4.48  54.5  244.
#> 10 OECD    2019  1.36  51.8  38.0  228. 12197.  4.40  53.6  236.
#> # ℹ 31 more rows
get_top_down_values(region_code = "PAK")
#> # A tibble: 41 × 10
#>    region    year     P     G     g     E     F     e     f    ef
#>    <chr>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Pakistan  2010 0.197 0.242  1.23  2.58  151. 10.7   58.5  623.
#>  2 Pakistan  2011 0.200 0.253  1.27  2.61  153. 10.3   58.6  604.
#>  3 Pakistan  2012 0.203 0.266  1.31  2.69  156. 10.1   58.2  587.
#>  4 Pakistan  2013 0.206 0.279  1.36  2.75  166.  9.87  60.4  597.
#>  5 Pakistan  2014 0.208 0.291  1.40  2.84  172.  9.75  60.7  592.
#>  6 Pakistan  2015 0.211 0.304  1.44  2.91  178.  9.59  61.1  586.
#>  7 Pakistan  2016 0.214 0.318  1.49  3.06  185.  9.63  60.6  583.
#>  8 Pakistan  2017 0.216 0.334  1.54  3.21  189.  9.62  58.8  565.
#>  9 Pakistan  2018 0.219 0.351  1.60  3.31  195.  9.43  58.9  555.
#> 10 Pakistan  2019 0.222 0.367  1.66  3.37  199.  9.18  59.0  541.
#> # ℹ 31 more rows