Get top-down projections of Kaya variables for one or more countries or regions

Get top-down projections of Kaya variables for one or more countries or regions

get_top_down_values(region_name, quiet = FALSE, region_code = NULL)

Arguments

region_name
The name of a country or region to look up

quiet
Suppress warnings if there is no data for that country or region.

region_code
Optional three-letter country or region code to look up instead of the region_name

Value

a tibble of values for P, G, E, F, g, e, f, and ef for each country or region:

region

The name of the country or region

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of constant 2010 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

See also

Examples

get_top_down_values("New Zealand")

#> # A tibble: 8 × 10
#>   region       year       P     G     g     E     F     e     f    ef
#>   <chr>       <int>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 New Zealand  2015 0.00460 0.169  36.8 0.843  36.5  4.98  43.3  216.
#> 2 New Zealand  2020 0.00496 0.193  38.8 0.879  36.0  4.57  41.0  187.
#> 3 New Zealand  2025 0.00534 0.219  41.0 0.953  37.7  4.35  39.6  172.
#> 4 New Zealand  2030 0.00568 0.248  43.6 1.03   39.8  4.14  38.8  160.
#> 5 New Zealand  2035 0.00603 0.282  46.8 1.11   42.0  3.94  37.8  149.
#> 6 New Zealand  2040 0.00636 0.320  50.3 1.21   44.7  3.78  37.0  140.
#> 7 New Zealand  2045 0.00669 0.362  54.1 1.32   47.9  3.65  36.3  132.
#> 8 New Zealand  2050 0.00700 0.409  58.4 1.45   51.9  3.55  35.7  127.

get_top_down_values("OECD")

#> # A tibble: 8 × 10
#>   region  year     P     G     g     E      F     e     f    ef
#>   <chr>  <int> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 OECD    2015  1.33  49.3  37.1  217. 12348.  4.40  57.0  250.
#> 2 OECD    2020  1.37  53.9  39.4  221. 12364.  4.09  56.1  230.
#> 3 OECD    2025  1.40  59.2  42.3  224. 12115.  3.78  54.1  205.
#> 4 OECD    2030  1.43  64.0  44.8  226. 11926.  3.54  52.7  186.
#> 5 OECD    2035  1.45  69.2  47.7  230. 12074.  3.33  52.4  174.
#> 6 OECD    2040  1.47  74.8  50.9  237. 12310.  3.16  52.0  165.
#> 7 OECD    2045  1.48  80.5  54.2  242. 12616.  3.01  52.0  157.
#> 8 OECD    2050  1.49  86.5  57.9  250. 13021.  2.89  52.1  150.

get_top_down_values(region_code = "PAK")

#> # A tibble: 8 × 10
#>   region    year     P     G     g     E     F     e     f    ef
#>   <chr>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Pakistan  2015 0.199 0.216  1.08  2.77  160. 12.8   57.7  742.
#> 2 Pakistan  2020 0.212 0.266  1.25  3.03  179. 11.4   59.2  673.
#> 3 Pakistan  2025 0.224 0.331  1.47  3.44  205. 10.4   59.6  619.
#> 4 Pakistan  2030 0.235 0.400  1.70  3.93  222.  9.82  56.4  554.
#> 5 Pakistan  2035 0.245 0.478  1.95  4.40  250.  9.21  56.7  523.
#> 6 Pakistan  2040 0.254 0.565  2.22  4.96  281.  8.78  56.7  498.
#> 7 Pakistan  2045 0.261 0.660  2.52  5.55  315.  8.42  56.7  477.
#> 8 Pakistan  2050 0.267 0.765  2.86  6.20  352.  8.11  56.9  461.