Get top-down projections of Kaya variables for a country or region

Get top-down projections of Kaya variables for a country or region

get_top_down_values(region_name, quiet = FALSE, region_code = NULL)

Arguments

region_name
The name of a country or region to look up

quiet
Suppress warnings if there is no data for that country or region.

region_code
Optional three-letter country or region code to look up instead of the `region_name`

Value

a tibble of values for P, G, E, F, g, e, f, and ef for the country or region:

region

The name of the country or region

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of constant 2010 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

Examples

get_top_down_values("New Zealand")

#> # A tibble: 8 x 10
#>   region       year       P     G     g     E     F     e     f    ef
#>   <chr>       <int>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 New Zealand  2015 0.00460 0.170  37.0 0.845  35.3  4.96  41.8  208.
#> 2 New Zealand  2020 0.00496 0.194  39.1 0.882  34.9  4.55  39.5  180.
#> 3 New Zealand  2025 0.00534 0.220  41.3 0.955  36.5  4.34  38.2  166.
#> 4 New Zealand  2030 0.00568 0.250  43.9 1.03   38.5  4.12  37.4  154.
#> 5 New Zealand  2035 0.00603 0.284  47.1 1.11   40.7  3.93  36.5  143.
#> 6 New Zealand  2040 0.00636 0.322  50.6 1.21   43.3  3.77  35.7  134.
#> 7 New Zealand  2045 0.00669 0.364  54.5 1.32   46.4  3.63  35.1  127.
#> 8 New Zealand  2050 0.00700 0.412  58.8 1.46   50.2  3.54  34.4  122.

get_top_down_values("OECD")

#> # A tibble: 8 x 10
#>   region  year     P     G     g     E      F     e     f    ef
#>   <chr>  <int> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 OECD    2015  1.28  48.8  38.0  218. 12411.  4.47  56.9  254.
#> 2 OECD    2020  1.32  53.3  40.3  222. 12427.  4.16  56.0  233.
#> 3 OECD    2025  1.35  58.6  43.3  225. 12177.  3.84  54.1  208.
#> 4 OECD    2030  1.38  63.3  45.9  227. 11987.  3.59  52.7  189.
#> 5 OECD    2035  1.40  68.5  48.8  232. 12136.  3.38  52.4  177.
#> 6 OECD    2040  1.42  74.0  52.1  238. 12373.  3.21  52.0  167.
#> 7 OECD    2045  1.44  79.7  55.5  244. 12681.  3.06  52.0  159.
#> 8 OECD    2050  1.45  85.7  59.2  251. 13088.  2.93  52.1  153.

get_top_down_values(region_code = "PAK")

#> # A tibble: 8 x 10
#>   region    year     P     G     g     E     F     e     f    ef
#>   <chr>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Pakistan  2015 0.189 0.216  1.14  2.79  161. 12.9   57.6  744.
#> 2 Pakistan  2020 0.202 0.267  1.32  3.05  180. 11.4   59.1  675.
#> 3 Pakistan  2025 0.213 0.331  1.55  3.46  206. 10.4   59.5  621.
#> 4 Pakistan  2030 0.224 0.401  1.79  3.96  223.  9.87  56.3  555.
#> 5 Pakistan  2035 0.233 0.478  2.05  4.43  251.  9.25  56.6  524.
#> 6 Pakistan  2040 0.241 0.565  2.35  4.99  282.  8.82  56.6  499.
#> 7 Pakistan  2045 0.248 0.660  2.66  5.59  316.  8.46  56.6  479.
#> 8 Pakistan  2050 0.254 0.766  3.02  6.24  354.  8.14  56.7  462.