kayadata package

kayadata is a package for working with Kaya identity data for many countries and regions.

The Kaya identity, named for the economist Yoichi Kaya, who introduced it (Kaya, 1998); It decomposes the energy-related carbon dioxide emissions from a nation, region, or the world into the product of four components: $$F = P \times g \times e \times f,$$ where F is the total emissions, P is the population, g is the per-capita GDP, e is the energy intensity of the economy, and f is the emissions-intensity of the energy supply. (Nakicenovic and Swart, 2000, Ch. 3, p. 105; Raupach et al, 2007)

The data in this packages covers 1960-2019 for population and GDP, and 1965-2019 for energy and fossil-fuel CO2 emissions.

The package uses data on population and GDP from the World Bank, using market exchange rates (MER) for GDP because those data go back to 1960. From 1990 onward, Purchasing-Power-Parity (PPP) GDP figures are available as G_ppp but using these would require re-calculating G, g, e, and ef in the kaya_data data frame.

The package uses data on energy consumption and fossil-fuel CO2 emissions from the Energy Institute's 2023 Statistical Review of World Energy

License

The kayadata package is open source licensed under the MIT License.

Bug reports

kayadata issue tracker (https://github.com/jonathan-g/kayadata/issues)

References

Kaya, Yoichi and Keiichi Yokobori, Environment, Energy, and Economy: Strategies for Sustainability (United Nations University Press, 1998).

Nakicenovic, Nebojsa, and Rob Swart (Eds.), Special Report on Emissions Scenarios (Cambridge University Press, 2000). https://www.ipcc.ch/report/emissions-scenarios/

Raupach, Michael R., et al., "Global and regional drivers of accelerating CO2 emissions," PNAS 104, 10288--10293 (2007) doi:10.1073/pnas.0700609104 .