Get top-down projections of Kaya variables for one or more countries or regions for a given year

Get top-down projections of Kaya variables for one or more countries or regions for a given year

project_top_down(region_name, year, quiet = FALSE, region_code = NULL)

Arguments

region_name
The name of a country or region to look up

year
The year to project to

quiet
Suppress warnings if there is no data for that country or region.

region_code
Optional three-letter country or region code to look up instead of the region_name

Value

a tibble of values for P, G, E, F, g, e, f, and ef for each country or region:

region

The name of the country or region

year

The year

P

Population, in billions

G

Gross domestic product, in trillions of constant 2010 U.S. dollars.

E

Total primary energy consumption, in quads

F

CO2 emissions from fossil fuel consumption, in millions of metric tons

g

Per-capita GDP, in thousands of constant 2010 U.S. dollars per person.

e

Energy intensity of the economy, in quads per trillion dollars.

f

Emissions intensity of the energy supply, in million metric tons per quad.

ef

Emissions intensity of the economy, in metric tons per million dollars of GDP.

See also

Examples

project_top_down("China", 2037)

#> # A tibble: 1 × 10
#>   region  year     P     G     g     E     F     e     f    ef
#>   <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 China   2037  1.41  23.3  16.6  151. 9482.  6.51  62.8  410.

project_top_down(region_code = "VNM", year = 2043)

#> # A tibble: 1 × 10
#>   region   year     P     G     g     E     F     e     f    ef
#>   <chr>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Vietnam  2043 0.120 0.445  3.71  5.27  346.  11.8  65.6  777.